+++ src/app/api-services/index.ts
@@ -4,3 +4,4 @@ export * from './api-service.module';
 export * from './preset.service';
 export * from './ro.service';
 export * from './summary.service';
+export * from './divine-pride.service';