+++ src/app/api-services/models/divine-pride-skill.model.ts
@@ -1,0 +1,11 @@
+export interface DivinePrideSkillModel {
+  globalization: DivinePrideSkillGlobalization[];
+  id: number;
+}
+
+export interface DivinePrideSkillGlobalization {
+  name: string;
+  description: string;
+  server: number;
+  language: number;
+}