+++ src/app/api-services/models/index.ts
@@ -14,3 +14,6 @@ export * from './published-presets.response';
 export * from './bulk-operation-tags.request';
 export * from './publish-preset-model';
 export * from './like-tag.response';
+export * from './divine-pride-skill.model';
+export * from './divine-pride-monster.model';
+export * from './divine-pride-item.model';